+++ spa/src/pages/Connections/GHEnterpriseConnections/GHEnterpriseApplication.tsx
@@ -1,16 +1,8 @@
 /** @jsxImportSource @emotion/react */
 import { DynamicTableStateless } from "@atlaskit/dynamic-table";
-import DropdownMenu, {
-	DropdownItem,
-	DropdownItemGroup,
-} from "@atlaskit/dropdown-menu";
 import { useState } from "react";
-import Button from "@atlaskit/button";
-import { Flex, xcss } from "@atlaskit/primitives";
-import MoreIcon from "@atlaskit/icon/glyph/more";
 import Heading from "@atlaskit/heading";
-import ChevronRightIcon from "@atlaskit/icon/glyph/chevron-right";
-import ChevronDownIcon from "@atlaskit/icon/glyph/chevron-down";
+import { css } from "@emotion/react";
 import {
 	head,
 	getGHSubscriptionsRows,
@@ -20,7 +12,7 @@ import {
 	GitHubEnterpriseApplication,
 	SuccessfulConnection,
 } from "../../../rest-interfaces";
-import { css } from "@emotion/react";
+import GHEnterpriseAppHeader from "./GHEnterpriseAppHeader";
 
 const connectNewAppLinkStyle = css`
 	text-decoration: none;
@@ -44,27 +36,15 @@ const noConnectionsBodyStyle = css`
 	padding-top: 10px;
 `;
 
-const applicationHeaderStyle = css`
-	cursor: pointer;
-	display: flex;
-	align-items: center;
-	justify-content: flex-start;
-	width: 100%;
-`;
-
 const applicationContentStyle = css`
 	width: 100%;
 `;
 
-const appHeaderContainerStyle = xcss({
-	width: "100%",
-	justifyContent: "space-between",
-	marginBottom: "20px",
-});
-
 type GitHubEnterpriseApplicationProps = {
 	application: GitHubEnterpriseApplication;
-	setDataForModal: (dataForModal: SuccessfulConnection | GitHubEnterpriseApplication) => void;
+	setDataForModal: (
+		dataForModal: SuccessfulConnection | GitHubEnterpriseApplication
+	) => void;
 	setSelectedModal: (selectedModal: BackfillPageModalTypes) => void;
 	setIsModalOpened: (isModalOpen: boolean) => void;
 };
@@ -86,8 +66,9 @@ const GitHubEnterpriseApp = ({
 	setDataForModal,
 	setSelectedModal,
 }: GitHubEnterpriseApplicationProps) => {
-
 	const [showAppContent, setShowAppContent] = useState<boolean>(true);
+	const toggleShowAppContent = () =>
+		setShowAppContent((prevState) => !prevState);
 	const onConnectNewApp = () => {
 		return AP.context.getToken((token: string) => {
 			const child: Window | null = openChildWindow(
@@ -99,55 +80,17 @@ const GitHubEnterpriseApp = ({
 			}
 		});
 	};
-	const onEditGitHubApp = () =>{
-		const uuid = application.uuid;
-		AP.navigator.go(
-			"addonmodule",
-			{
-				moduleKey: "github-edit-app-page",
-				customData: { uuid }
-			}
-		);
-	};
+
 	return (
 		<div css={wrapperStyle}>
-			<Flex xcss={appHeaderContainerStyle}>
-				<div
-					css={applicationHeaderStyle}
-					onClick={() => {
-						setShowAppContent((prevState) => !prevState);
-					}}
-				>
-					{showAppContent ? (
-						<ChevronDownIcon label="" />
-					) : (
-						<ChevronRightIcon label="" />
-					)}
-					<Heading level="h400">{application.gitHubAppName}</Heading>
-				</div>
-				<div>
-					<DropdownMenu
-						trigger={({ triggerRef, ...props }) => (
-							<Button
-								{...props}
-								appearance="subtle"
-								iconBefore={<MoreIcon label="more" size="small" />}
-								ref={triggerRef}
-							/>
-						)}
-					>
-						<DropdownItemGroup>
-							<DropdownItem onClick={onEditGitHubApp}>Edit</DropdownItem>
-							<DropdownItem onClick={()=>{
-								setIsModalOpened(true);
-								setDataForModal(application);
-								setSelectedModal("DISCONNECT_SERVER_APP");
-							}}>Disconnect</DropdownItem>
-						</DropdownItemGroup>
-					</DropdownMenu>
-				</div>
-			</Flex>
-
+			<GHEnterpriseAppHeader
+				application={application}
+				setDataForModal={setDataForModal}
+				setIsModalOpened={setIsModalOpened}
+				setSelectedModal={setSelectedModal}
+				showAppContent={showAppContent}
+				toggleShowAppContent={toggleShowAppContent}
+			/>
 			{showAppContent && (
 				<>
 					{application.successfulConnections.length > 0 ? (
